upgrade go-test, dependencies, and use math/rand/v2
The new go-test/random package supports a reusable random number generator and removes support for a global seed, which can lead to tests interferring with eachother.

+++ testutil/testchain.go
@@ -72,7 +72,7 @@ func createBlock(parents []ipld.Link, size uint64) (ipld.Node, error) {
 	if err != nil {
 		return nil, err
 	}
-	err = mnblnb.AssembleValue().AssignBytes(random.Bytes(int(size)))
+	err = mnblnb.AssembleValue().AssignBytes(random.Bytes(int64(size)))
 	if err != nil {
 		return nil, err
 	}
